refactor(folio): move agnostic paged-layout logic into core
Relocate the 14 framework-agnostic paged-editor modules (measurement,
geometry, scroll, selection-gate, table-resize math) and their unit tests
into core/paged-layout/. paged-editor/ now holds only the React rendering
(.tsx) and the 2 React hooks, so the headless core owns the full paged-layout
pipeline and a non-React adapter can drive it. Pure relocation plus import
repointing; no logic changes.
