+++ packages/folio/src/core/paged-layout/LayoutSelectionGate.test.ts
