+++ packages/folio/src/core/paged-layout/LayoutSelectionGate.ts
