+++ packages/folio/src/core/paged-layout/hiddenEditorScroll.ts
