+++ packages/folio/src/core/paged-layout/readOnlyEditAttempt.ts
