+++ packages/folio/src/core/paged-layout/selectionViewportRect.ts
