+++ packages/folio/src/core/paged-layout/tableColumnResize.ts
