+++ packages/folio/src/core/paged-layout/transactionDirtyRange.ts
